+++ packages/nimble-components/src/rich-text-editor/styles.ts
@@ -14,7 +14,6 @@ import {
     smallDelay,
     standardPadding
 } from '../theme-provider/design-tokens';
-import { userSelectNone } from '../utilities/style/user-select';
 import { styles as errorStyles } from '../patterns/error/styles';
 
 export const styles = css`
@@ -30,7 +29,7 @@ export const styles = css`
             ${borderWidth} + 1px
         );
         --ni-private-rich-text-editor-footer-section-height: 40px;
-        --ni-private-rich-text-editor-editor-minimum-height: 32px;
+        --ni-private-rich-text-editor-tiptap-editor-minimum-height: 36px;
         ${
             /** Minimum width is added to accommodate all the possible buttons in the toolbar and to support the mobile width. */ ''
         }
@@ -69,27 +68,29 @@ export const styles = css`
         }
     }
 
+    :host([disabled]) .container {
+        color: ${bodyDisabledFontColor};
+        border: ${borderWidth} solid rgba(${borderRgbPartialColor}, 0.1);
+    }
+
+    :host([error-visible]) .container {
+        border-bottom-color: ${failColor};
+    }
+
     :host(:hover) .container::after {
         width: 100%;
-    }
-
-    :host([disabled]) *,
-    :host([disabled]) {
-        ${userSelectNone}
-        color: ${bodyDisabledFontColor};
-    }
-
-    :host([disabled]) .container,
-    :host([disabled]) .container::after {
-        border: ${borderWidth} solid rgba(${borderRgbPartialColor}, 0.1);
     }
 
     :host([disabled]:hover) .container::after {
         width: 0px;
     }
 
-    :host([disabled]) .ProseMirror p.is-editor-empty:first-child::before {
-        color: ${controlLabelDisabledFontColor};
+    :host([error-visible]) .container::after {
+        border-bottom-color: ${failColor};
+    }
+
+    .editor-container {
+        display: contents;
     }
 
     .editor {
@@ -101,10 +102,6 @@ export const styles = css`
         overflow: auto;
     }
 
-    .editor-container {
-        display: contents;
-    }
-
     .ProseMirror {
         ${
             /**
@@ -112,7 +109,7 @@ export const styles = css`
              * However, max height will be `fit-content` when the `fit-to-content` attribute for the editor component is implemented.
              */ ''
         }
-        min-height: var(--ni-private-rich-text-editor-editor-minimum-height);
+        min-height: var(--ni-private-rich-text-editor-tiptap-editor-minimum-height);
         max-height: 132px;
         height: 100%;
         border: ${borderWidth} solid transparent;
@@ -167,8 +164,38 @@ export const styles = css`
         margin-block-end: 0;
     }
 
+    :host([fit-to-content]) .ProseMirror {
+        max-height: fit-content;
+    }
+
+    :host([footer-hidden]) .ProseMirror {
+        ${
+            /**
+             * Minimum height is the addition of existing minimum height of the tiptap editor div, the footer section height and the top and bottom border width.
+             * With this calculation, the editor will extend to use the footer height when it is hidden.
+             *
+             * Use case: If the footer is initially hidden and is dynamically enabled when the editor is focused, there will be no layout shift,
+             * and the footer will smoothly appear within the editor.
+             */ ''
+        }
+        min-height: calc(var(--ni-private-rich-text-editor-tiptap-editor-minimum-height) + var(--ni-private-rich-text-editor-footer-section-height) + calc(${borderWidth} * 2));
+    }
+
     li > p {
         margin-block: 0;
+    }
+
+    .ProseMirror p.is-editor-empty:first-child::before {
+        color: ${controlLabelFontColor};
+        content: attr(data-placeholder);
+        float: left;
+        height: 0;
+        pointer-events: none;
+        word-break: break-word;
+    }
+
+    :host([disabled]) .ProseMirror p.is-editor-empty:first-child::before {
+        color: ${controlLabelDisabledFontColor};
     }
 
     .footer-section {
@@ -182,25 +209,7 @@ export const styles = css`
     }
 
     :host([footer-hidden]) .footer-section {
-        visibility: hidden;
-        height: 0px;
-    }
-
-    :host([footer-hidden]) .ProseMirror {
-        ${
-            /**
-             * Minimum height is the combination of existing minimum height of the tiptap div, footer height and extra padding added to the tiptap div.
-             * With this calculation, the editor will extend to use the footer height when it is hidden.
-             *
-             * Use case: If the footer is initially hidden and is dynamically enabled when the editor is focused, there will be no layout shift,
-             * and the footer will smoothly appear within the editor.
-             */ ''
-        }
-        min-height: calc(var(--ni-private-rich-text-editor-editor-minimum-height) + var(--ni-private-rich-text-editor-footer-section-height) + 4px);
-    }
-
-    :host([fit-to-content]) .ProseMirror {
-        max-height: fit-content;
+        display: none;
     }
 
     nimble-toolbar::part(positioning-region) {
@@ -220,29 +229,7 @@ export const styles = css`
         place-items: center;
     }
 
-    .ProseMirror p.is-editor-empty:first-child::before {
-        color: ${controlLabelFontColor};
-        content: attr(data-placeholder);
-        float: left;
-        height: 0;
-        pointer-events: none;
-        word-break: break-all;
-    }
-
     :host([error-visible]) .error-icon {
-        display: none;
-    }
-
-    :host([error-visible]) .container {
-        border-bottom-color: ${failColor};
-    }
-
-    :host([error-visible]) .container::after {
-        border-bottom-color: ${failColor};
-    }
-
-    :host([error-visible]) .error-icon.scrollbar-width-calculated {
-        display: inline-flex;
         position: absolute;
         top: calc(${standardPadding} / 2);
         right: var(--ni-private-rich-text-editor-scrollbar-width);